services/indexeur-semantique/requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Embeddings & Vector Search
sentence-transformers==2.2.2
faiss-cpu==1.7.4
# faiss-gpu==1.7.4  # Uncomment for GPU support
transformers==4.35.2
torch==2.1.1

# Medical Embeddings
# Note: Install medical-specific models at runtime

# Text Processing
nltk==3.8.1
spacy==3.7.2

# Message Queue
pika==1.3.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
numpy==1.26.2

# Logging
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
